requirements.txt
bottle==0.12.25 # web ui
pycryptodomex==3.18.0 # encryption
